+++ Mintegral/src/main/java/com/mopub/mobileads/MintegralInterstitial.java
@@ -4,11 +4,16 @@ import android.app.Activity;
 import android.content.Context;
 import android.text.TextUtils;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 import com.mintegral.msdk.MIntegralConstans;
 import com.mintegral.msdk.interstitialvideo.out.InterstitialVideoListener;
 import com.mintegral.msdk.interstitialvideo.out.MTGBidInterstitialVideoHandler;
 import com.mintegral.msdk.interstitialvideo.out.MTGInterstitialVideoHandler;
 import com.mintegral.msdk.out.MIntegralSDKFactory;
+import com.mopub.common.LifecycleListener;
+import com.mopub.common.Preconditions;
 import com.mopub.common.logging.MoPubLog;
 
 import java.util.Map;
@@ -29,29 +34,28 @@ import static com.mopub.mobileads.MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR;
 import static com.mopub.mobileads.MoPubErrorCode.NETWORK_NO_FILL;
 import static com.mopub.mobileads.MoPubErrorCode.UNSPECIFIED;
 
-public class MintegralInterstitial extends CustomEventInterstitial implements InterstitialVideoListener {
+public class MintegralInterstitial extends BaseAd implements InterstitialVideoListener {
 
     private final String ADAPTER_NAME = this.getClass().getSimpleName();
 
     private MTGInterstitialVideoHandler mInterstitialHandler;
     private MTGBidInterstitialVideoHandler mBidInterstitialVideoHandler;
-    private CustomEventInterstitialListener mCustomEventInterstitialListener;
 
-    private static String mAdUnitId;
-    private static String mPlacementId;
+    private String mAdUnitId;
+    private String mPlacementId;
 
     @Override
-    protected void loadInterstitial(final Context context,
-                                    final CustomEventInterstitialListener customEventInterstitialListener,
-                                    final Map<String, Object> localExtras, Map<String, String> serverExtras) {
+    protected void load(@NonNull final Context context, @NonNull final AdData adData) {
+        Preconditions.checkNotNull(context);
+        Preconditions.checkNotNull(adData);
 
         setAutomaticImpressionAndClickTracking(false);
-        mCustomEventInterstitialListener = customEventInterstitialListener;
 
-        if (!serverDataIsValid(serverExtras, context)) {
+        final Map<String, String> extras = adData.getExtras();
+        if (!serverDataIsValid(extras, context)) {
             failAdapter(LOAD_FAILED, ADAPTER_CONFIGURATION_ERROR, "One or " +
                     "more keys used for Mintegral's ad requests are empty. Failing adapter. Please " +
-                    "ensure you have populated all the required keys on the MoPub dashboard.");
+                    "ensure you have populated all the required keys on the MoPub dashboard.", true);
 
             return;
         }
@@ -60,34 +64,19 @@ public class MintegralInterstitial extends CustomEventInterstitial implements In
         MintegralAdapterConfiguration.setTargeting(MIntegralSDKFactory.getMIntegralSDK());
 
         if (context instanceof Activity) {
-            final String adm = serverExtras.get(ADM_KEY);
+            final String adMarkup = extras.get(ADM_KEY);
 
-            if (TextUtils.isEmpty(adm)) {
-                mInterstitialHandler = MintegralHandlerManager.getInstance()
-                        .getMTGInterstitialVideoHandler(mAdUnitId);
-
-                if (mInterstitialHandler == null) {
-                    mInterstitialHandler = new MTGInterstitialVideoHandler(context, mPlacementId,
-                            mAdUnitId);
-                    MintegralHandlerManager.getInstance().addMTGInterstitialVideoHandler(mAdUnitId,
-                            mInterstitialHandler);
-                }
+            if (TextUtils.isEmpty(adMarkup)) {
+                mInterstitialHandler = new MTGInterstitialVideoHandler(context, mPlacementId, mAdUnitId);
                 mInterstitialHandler.setRewardVideoListener(this);
                 mInterstitialHandler.load();
 
                 handleAudio();
             } else {
-                mBidInterstitialVideoHandler = MintegralHandlerManager.getInstance()
-                        .getMTGBidInterstitialVideoHandler(mAdUnitId);
-
-                if (mBidInterstitialVideoHandler == null) {
-                    mBidInterstitialVideoHandler = new MTGBidInterstitialVideoHandler(context, mPlacementId,
-                            mAdUnitId);
-                    MintegralHandlerManager.getInstance().addMTGBidInterstitialVideoHandler(mAdUnitId,
-                            mBidInterstitialVideoHandler);
-                }
+                mBidInterstitialVideoHandler = new MTGBidInterstitialVideoHandler(context,
+                        mPlacementId, mAdUnitId);
                 mBidInterstitialVideoHandler.setRewardVideoListener(this);
-                mBidInterstitialVideoHandler.loadFromBid(adm);
+                mBidInterstitialVideoHandler.loadFromBid(adMarkup);
 
                 handleAudio();
             }
@@ -95,12 +84,12 @@ public class MintegralInterstitial extends CustomEventInterstitial implements In
             MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
         } else {
             failAdapter(LOAD_FAILED, ADAPTER_CONFIGURATION_ERROR, "Context is not an instance " +
-                    "of Activity. Aborting ad request, and failing adapter.");
+                    "of Activity. Aborting ad request, and failing adapter.", true);
         }
     }
 
     @Override
-    protected void showInterstitial() {
+    protected void show() {
         if (mInterstitialHandler != null && mInterstitialHandler.isReady()) {
             handleAudio();
             mInterstitialHandler.show();
@@ -109,7 +98,7 @@ public class MintegralInterstitial extends CustomEventInterstitial implements In
             mBidInterstitialVideoHandler.showFromBid();
         } else {
             failAdapter(SHOW_FAILED, NETWORK_NO_FILL, "Failed to show Mintegral interstitial " +
-                    "because it is not ready. Please make a new ad request.");
+                    "because it is not ready. Please make a new ad request.", false);
         }
 
         MoPubLog.log(getAdNetworkId(), SHOW_ATTEMPTED, ADAPTER_NAME);
@@ -129,12 +118,16 @@ public class MintegralInterstitial extends CustomEventInterstitial implements In
             mBidInterstitialVideoHandler.setInterstitialVideoListener(null);
             mBidInterstitialVideoHandler = null;
         }
+    }
 
-        mCustomEventInterstitialListener = null;
+    @Nullable
+    @Override
+    protected LifecycleListener getLifecycleListener() {
+        return null;
     }
 
     private void failAdapter(final MoPubLog.AdapterLogEvent event, final MoPubErrorCode errorCode,
-                             final String errorMsg) {
+                             final String errorMsg, final boolean isLoad) {
 
         MoPubLog.log(getAdNetworkId(), event, ADAPTER_NAME, errorCode.getIntCode(), errorCode);
 
@@ -142,14 +135,16 @@ public class MintegralInterstitial extends CustomEventInterstitial implements In
             MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, errorMsg);
         }
 
-        if (mCustomEventInterstitialListener != null) {
-            mCustomEventInterstitialListener.onInterstitialFailed(errorCode);
+        if (isLoad && mLoadListener != null) {
+            mLoadListener.onAdLoadFailed(errorCode);
+        } else if (!isLoad && mInteractionListener != null) {
+            mInteractionListener.onAdFailed(errorCode);
         }
     }
 
     private void handleAudio() {
-        boolean isMute = MintegralAdapterConfiguration.isMute();
-        int muteStatus = isMute ? REWARD_VIDEO_PLAY_MUTE : REWARD_VIDEO_PLAY_NOT_MUTE;
+        final boolean isMute = MintegralAdapterConfiguration.isMute();
+        final int muteStatus = isMute ? REWARD_VIDEO_PLAY_MUTE : REWARD_VIDEO_PLAY_NOT_MUTE;
 
         if (mInterstitialHandler != null) {
             mInterstitialHandler.playVideoMute(muteStatus);
@@ -158,14 +153,14 @@ public class MintegralInterstitial extends CustomEventInterstitial implements In
         }
     }
 
-    private boolean serverDataIsValid(final Map<String, String> serverExtras, Context context) {
+    private boolean serverDataIsValid(final Map<String, String> extras, Context context) {
 
-        if (serverExtras != null && !serverExtras.isEmpty()) {
-            mAdUnitId = serverExtras.get(MintegralAdapterConfiguration.UNIT_ID_KEY);
-            mPlacementId = serverExtras.get(MintegralAdapterConfiguration.PLACEMENT_ID_KEY);
+        if (extras != null && !extras.isEmpty()) {
+            mAdUnitId = extras.get(MintegralAdapterConfiguration.UNIT_ID_KEY);
+            mPlacementId = extras.get(MintegralAdapterConfiguration.PLACEMENT_ID_KEY);
 
-            final String appId = serverExtras.get(MintegralAdapterConfiguration.APP_ID_KEY);
-            final String appKey = serverExtras.get(MintegralAdapterConfiguration.APP_KEY);
+            final String appId = extras.get(MintegralAdapterConfiguration.APP_ID_KEY);
+            final String appKey = extras.get(MintegralAdapterConfiguration.APP_KEY);
 
             if (!TextUtils.isEmpty(appId) && !TextUtils.isEmpty(appKey) && !TextUtils.isEmpty(mAdUnitId)) {
                 MintegralAdapterConfiguration.configureMintegral(appId, appKey, context);
@@ -176,38 +171,45 @@ public class MintegralInterstitial extends CustomEventInterstitial implements In
         return false;
     }
 
-    private static String getAdNetworkId() {
-        return mAdUnitId;
+    @NonNull
+    protected String getAdNetworkId() {
+        return mAdUnitId != null ? mAdUnitId : "";
+    }
+
+    @Override
+    protected boolean checkAndInitializeSdk(@NonNull final Activity activity,
+                                            @NonNull final AdData adData) {
+        return false;
     }
 
     @Override
     public void onVideoLoadSuccess(String placementId, String s) {
         MoPubLog.log(getAdNetworkId(), LOAD_SUCCESS, ADAPTER_NAME);
 
-        if (mCustomEventInterstitialListener != null) {
-            mCustomEventInterstitialListener.onInterstitialLoaded();
+        if (mLoadListener != null) {
+            mLoadListener.onAdLoaded();
         }
     }
 
     @Override
     public void onVideoLoadFail(String errorMsg) {
-        failAdapter(LOAD_FAILED, UNSPECIFIED, errorMsg);
+        failAdapter(LOAD_FAILED, UNSPECIFIED, errorMsg, true);
     }
 
     @Override
     public void onAdShow() {
         MoPubLog.log(getAdNetworkId(), SHOW_SUCCESS, ADAPTER_NAME);
 
-        if (mCustomEventInterstitialListener != null) {
-            mCustomEventInterstitialListener.onInterstitialShown();
-            mCustomEventInterstitialListener.onInterstitialImpression();
+        if (mInteractionListener != null) {
+            mInteractionListener.onAdShown();
+            mInteractionListener.onAdImpression();
         }
     }
 
     @Override
     public void onShowFail(String errorMsg) {
         failAdapter(SHOW_FAILED, UNSPECIFIED, "Failed to show Mintegral interstitial: "
-                + errorMsg);
+                + errorMsg, false);
     }
 
     @Override
@@ -215,8 +217,8 @@ public class MintegralInterstitial extends CustomEventInterstitial implements In
         MoPubLog.log(getAdNetworkId(), DID_DISAPPEAR, ADAPTER_NAME);
         MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "onAdClose");
 
-        if (mCustomEventInterstitialListener != null) {
-            mCustomEventInterstitialListener.onInterstitialDismissed();
+        if (mInteractionListener != null) {
+            mInteractionListener.onAdDismissed();
         }
     }
 
@@ -224,8 +226,8 @@ public class MintegralInterstitial extends CustomEventInterstitial implements In
     public void onVideoAdClicked(String placementId, String message) {
         MoPubLog.log(getAdNetworkId(), CLICKED, ADAPTER_NAME);
 
-        if (mCustomEventInterstitialListener != null) {
-            mCustomEventInterstitialListener.onInterstitialClicked();
+        if (mInteractionListener != null) {
+            mInteractionListener.onAdClicked();
         }
     }