+++ Mintegral/src/main/java/com/mopub/mobileads/MintegralRewardedVideo.java
@@ -36,7 +36,7 @@ import static com.mopub.mobileads.MoPubErrorCode.NETWORK_NO_FILL;
 import static com.mopub.mobileads.MoPubErrorCode.UNSPECIFIED;
 import static com.mopub.mobileads.MoPubErrorCode.VIDEO_PLAYBACK_ERROR;
 
-public class MintegralRewardedVideo extends CustomEventRewardedVideo implements RewardVideoListener {
+public class MintegralRewardedVideo extends BaseAd implements RewardVideoListener {
 
     private final String ADAPTER_NAME = this.getClass().getSimpleName();
 
@@ -64,18 +64,15 @@ public class MintegralRewardedVideo extends CustomEventRewardedVideo implements 
 
     @Override
     protected boolean checkAndInitializeSdk(@NonNull final Activity launcherActivity,
-                                            @NonNull final Map<String, Object> localExtras,
-                                            @NonNull final Map<String, String> serverExtras) throws Exception {
-
+                                            @NonNull final AdData adData) {
         Preconditions.checkNotNull(launcherActivity);
-        Preconditions.checkNotNull(localExtras);
-        Preconditions.checkNotNull(serverExtras);
+        Preconditions.checkNotNull(adData);
 
         mContext = launcherActivity.getApplicationContext();
         mUserId = MintegralAdapterConfiguration.getUserId();
         mRewardId = MintegralAdapterConfiguration.getRewardId();
 
-        if (!serverDataIsValid(serverExtras, mContext)) {
+        if (!serverDataIsValid(adData.getExtras(), mContext)) {
             failAdapter(LOAD_FAILED, ADAPTER_CONFIGURATION_ERROR, "One or " +
                     "more keys used for Mintegral's ad requests are empty. Failing adapter. " +
                     "Please ensure you have populated all the required keys on the MoPub " +
@@ -88,15 +85,15 @@ public class MintegralRewardedVideo extends CustomEventRewardedVideo implements 
     }
 
     @Override
-    protected void loadWithSdkInitialized(@NonNull final Activity activity,
-                                          @NonNull final Map<String, Object> localExtras,
-                                          @NonNull final Map<String, String> serverExtras) throws Exception {
+    protected void load(@NonNull final Context context,
+                        @NonNull final AdData adData) {
+        Preconditions.checkNotNull(context);
+        Preconditions.checkNotNull(adData);
 
-        Preconditions.checkNotNull(activity);
-        Preconditions.checkNotNull(localExtras);
-        Preconditions.checkNotNull(serverExtras);
+        setAutomaticImpressionAndClickTracking(false);
 
-        if (!serverDataIsValid(serverExtras, mContext)) {
+        final Map<String, String> extras = adData.getExtras();
+        if (!serverDataIsValid(extras, mContext)) {
             failAdapter(LOAD_FAILED, ADAPTER_CONFIGURATION_ERROR, "One or " +
                             "more keys used for Mintegral's ad requests are empty. Failing adapter. Please " +
                             "ensure you have populated all the required keys on the MoPub dashboard",
@@ -108,32 +105,18 @@ public class MintegralRewardedVideo extends CustomEventRewardedVideo implements 
         MintegralAdapterConfiguration.addChannel();
         MintegralAdapterConfiguration.setTargeting(MIntegralSDKFactory.getMIntegralSDK());
 
-        final String adm = serverExtras.get(ADM_KEY);
+        final String adMarkup = extras.get(ADM_KEY);
 
-        if (TextUtils.isEmpty(adm)) {
-            mMtgRewardVideoHandler = MintegralHandlerManager.getInstance().getMTGRewardVideoHandler(mAdUnitId);
-
-            if (mMtgRewardVideoHandler == null) {
-                mMtgRewardVideoHandler = new MTGRewardVideoHandler(mPlacementId, mAdUnitId);
-                MintegralHandlerManager.getInstance().addMTGRewardVideoHandler(mAdUnitId,
-                        mMtgRewardVideoHandler);
-            }
-
+        if (TextUtils.isEmpty(adMarkup)) {
+            mMtgRewardVideoHandler = new MTGRewardVideoHandler(mPlacementId, mAdUnitId);
             mMtgRewardVideoHandler.setRewardVideoListener(this);
             mMtgRewardVideoHandler.load();
 
             handleAudio();
         } else {
-            mtgBidRewardVideoHandler = MintegralHandlerManager.getInstance().getMTGBidRewardVideoHandler(mAdUnitId);
-
-            if (mtgBidRewardVideoHandler == null) {
-                mtgBidRewardVideoHandler = new MTGBidRewardVideoHandler(mPlacementId, mAdUnitId);
-                MintegralHandlerManager.getInstance().addMTGBidRewardVideoHandler(mAdUnitId,
-                        mtgBidRewardVideoHandler);
-            }
-
+            mtgBidRewardVideoHandler = new MTGBidRewardVideoHandler(mPlacementId, mAdUnitId);
             mtgBidRewardVideoHandler.setRewardVideoListener(this);
-            mtgBidRewardVideoHandler.loadFromBid(adm);
+            mtgBidRewardVideoHandler.loadFromBid(adMarkup);
 
             handleAudio();
         }
@@ -142,7 +125,7 @@ public class MintegralRewardedVideo extends CustomEventRewardedVideo implements 
     }
 
     @Override
-    protected void showVideo() {
+    protected void show() {
 
         if (mMtgRewardVideoHandler != null && mMtgRewardVideoHandler.isReady()) {
             handleAudio();
@@ -176,27 +159,14 @@ public class MintegralRewardedVideo extends CustomEventRewardedVideo implements 
         }
     }
 
-    @Override
-    protected boolean hasVideoAvailable() {
-        if (mMtgRewardVideoHandler != null) {
-            return mMtgRewardVideoHandler.isReady();
-        }
+    private boolean serverDataIsValid(final Map<String, String> extras, Context context) {
 
-        if (mtgBidRewardVideoHandler != null) {
-            return mtgBidRewardVideoHandler.isBidReady();
-        }
+        if (extras != null && !extras.isEmpty()) {
+            mAdUnitId = extras.get(MintegralAdapterConfiguration.UNIT_ID_KEY);
+            mPlacementId = extras.get(MintegralAdapterConfiguration.PLACEMENT_ID_KEY);
 
-        return false;
-    }
-
-    private boolean serverDataIsValid(final Map<String, String> serverExtras, Context context) {
-
-        if (serverExtras != null && !serverExtras.isEmpty()) {
-            mAdUnitId = serverExtras.get(MintegralAdapterConfiguration.UNIT_ID_KEY);
-            mPlacementId = serverExtras.get(MintegralAdapterConfiguration.PLACEMENT_ID_KEY);
-
-            final String appId = serverExtras.get(MintegralAdapterConfiguration.APP_ID_KEY);
-            final String appKey = serverExtras.get(MintegralAdapterConfiguration.APP_KEY);
+            final String appId = extras.get(MintegralAdapterConfiguration.APP_ID_KEY);
+            final String appKey = extras.get(MintegralAdapterConfiguration.APP_KEY);
 
             if (!TextUtils.isEmpty(appId) && !TextUtils.isEmpty(appKey) && !TextUtils.isEmpty(mAdUnitId)) {
                 if (!isInitialized) {
@@ -219,12 +189,10 @@ public class MintegralRewardedVideo extends CustomEventRewardedVideo implements 
             MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, errorMsg);
         }
 
-        if (loadRelated) {
-            MoPubRewardedVideoManager.onRewardedVideoLoadFailure(this.getClass(),
-                    mAdUnitId, errorCode);
-        } else {
-            MoPubRewardedVideoManager.onRewardedVideoPlaybackError(this.getClass(),
-                    mAdUnitId, errorCode);
+        if (loadRelated && mLoadListener != null) {
+            mLoadListener.onAdLoadFailed(errorCode);
+        } else if (!loadRelated && mInteractionListener != null) {
+            mInteractionListener.onAdFailed(errorCode);
         }
     }
 
@@ -242,19 +210,24 @@ public class MintegralRewardedVideo extends CustomEventRewardedVideo implements 
     @Override
     public void onAdClose(boolean b, String label, float amount) {
         if (b) {
-            MoPubRewardedVideoManager.onRewardedVideoCompleted(this.getClass(), mAdUnitId,
-                    MoPubReward.success(label, (int) amount));
+            if (mInteractionListener != null) {
+                mInteractionListener.onAdComplete(MoPubReward.success(label, (int) amount));
+            }
 
             MoPubLog.log(getAdNetworkId(), SHOULD_REWARD, ADAPTER_NAME, amount, label);
         }
 
-        MoPubRewardedVideoManager.onRewardedVideoClosed(this.getClass(), mAdUnitId);
+        if (mInteractionListener != null) {
+            mInteractionListener.onAdDismissed();
+        }
         MoPubLog.log(getAdNetworkId(), DID_DISAPPEAR, ADAPTER_NAME);
     }
 
     @Override
     public void onVideoLoadSuccess(String placementId, String unitId) {
-        MoPubRewardedVideoManager.onRewardedVideoLoadSuccess(this.getClass(), mAdUnitId);
+        if (mLoadListener != null) {
+            mLoadListener.onAdLoaded();
+        }
         MoPubLog.log(getAdNetworkId(), LOAD_SUCCESS, ADAPTER_NAME);
     }
 
@@ -272,7 +245,10 @@ public class MintegralRewardedVideo extends CustomEventRewardedVideo implements 
 
     @Override
     public void onAdShow() {
-        MoPubRewardedVideoManager.onRewardedVideoStarted(this.getClass(), mAdUnitId);
+        if (mInteractionListener != null) {
+            mInteractionListener.onAdShown();
+            mInteractionListener.onAdImpression();
+        }
         MoPubLog.log(getAdNetworkId(), SHOW_SUCCESS, ADAPTER_NAME);
     }
 
@@ -283,7 +259,9 @@ public class MintegralRewardedVideo extends CustomEventRewardedVideo implements 
 
     @Override
     public void onVideoAdClicked(String placementId, String unitId) {
-        MoPubRewardedVideoManager.onRewardedVideoClicked(this.getClass(), mAdUnitId);
+        if (mInteractionListener != null) {
+            mInteractionListener.onAdClicked();
+        }
         MoPubLog.log(getAdNetworkId(), CLICKED, ADAPTER_NAME);
     }